Moved Entities to separate folders
