+++ src/Entities/Task/types.ts
@@ -1,0 +1,10 @@
+import { Optional } from 'sequelize';
+
+export interface ITask {
+  id: string;
+  title: string;
+  description?: string;
+  userId: string;
+}
+
+export type ITaskCreationAttributes = Optional<ITask, 'id' | 'description'>;
