+++ src/Entities/User/types.ts
@@ -6,13 +6,4 @@ export interface IUser {
   password: string;
 }
 
-export interface ITask {
-  id: string;
-  title: string;
-  description?: string;
-  userId: string;
-}
-
-export type ITaskCreationAttributes = Optional<ITask, 'id' | 'description'>;
-
 export type IUserCreationAttributes = Optional<IUser, 'id'>;
